requirements.txt
Django==2.2
amqp>=1.4.6
anyjson>=0.3.3
billiard>=3.6.3.0
celery==4.4.2
kombu>=4.6.8
pytz>=2015.4
redis>=3.2.0
requests==2.7.0
psycopg2-binary==2.8.6
asgiref==3.3.4
django-cors-headers==3.0.0
djangorestframework==3.12.4
sqlparse==0.4.1